fix/useChatConnect
